snapshot of project "lynx", label v2-8-8dev_5b

+++ WWW/Library/Implementation/HTAtom.c
@@ -40,7 +40,7 @@ static void free_atoms(void);
 
 HTAtom *HTAtom_for(const char *string)
 {
-    int hash;
+    size_t hash;
     HTAtom *a;
 
     /* First time around, clear hash table
